+++ tests/test_extra_links.py
@@ -7,7 +7,7 @@ from sphinx_testing import with_app
 
 
 @with_app(buildername='html', srcdir='doc_test/doc_extra_links')
-def test_extra_links(app, status, warning):
+def test_extra_links_html(app, status, warning):
     app.build()
     html = Path(app.outdir, 'index.html').read_text()
     assert 'TEST_001' in html
@@ -16,4 +16,14 @@ def test_extra_links(app, status, warning):
     assert 'blocked by' in html
     assert 'blocks' in html
 
+
+@with_app(buildername='latex', srcdir='doc_test/doc_extra_links')
+def test_extra_links_latex(app, status, warning):
+    app.build()
+    tex = Path(app.outdir, 'needstestdocs.tex').read_text()
+    assert 'TEST_001' in tex
+    assert 'tested by' in tex
+    assert 'tests' in tex
+    assert 'blocked by' in tex
+    assert 'blocks' in tex